moves rate calculation to its own file
