+++ tesseract-clickhouse/src/sql.rs
@@ -1,10 +1,9 @@
-use itertools::join;
-
 mod aggregator;
 mod cuts;
 mod growth;
 mod options;
 mod primary_agg;
+mod rate;
 mod rca;
 
 use tesseract_core::query_ir::{
@@ -24,6 +23,7 @@ use tesseract_core::query_ir::{
 };
 use self::options::wrap_options;
 use self::primary_agg::primary_agg;
+use self::rate::rate_calculation;
 
 
 /// Error checking is done before this point. This string formatter
@@ -46,9 +46,9 @@ pub fn clickhouse_sql(
 {
     let (mut final_sql, mut final_drill_cols) = {
         if let Some(rca) = rca {
-            rca::calculate(table, cuts, drills, meas, rate, rca)
+            rca::calculate(table, cuts, drills, meas, rca)
         } else {
-            primary_agg(table, cuts, drills, meas, rate)
+            primary_agg(table, cuts, drills, meas)
         }
     };
 
@@ -68,67 +68,6 @@ pub fn clickhouse_sql(
 
     final_sql
 }
-
-
-pub fn rate_calculation(
-    table: &TableSql,
-    cuts: &[CutSql],
-    drills: &[DrilldownSql],
-    meas: &[MeasureSql],
-    rate: &RateSql,
-    final_sql: &str,
-    final_drill_cols: &str
-) -> String
-{
-    let final_sql = final_sql.to_string().clone();
-    let final_drill_cols = final_drill_cols.to_string().clone();
-
-    let mut rate_sql = "(select ".to_string();
-    let mut drill_aliases: Vec<String> = vec![];
-
-    for drill in drills {
-        let drill_alias = &drill.col_alias_only_vec()[0];
-
-        drill_aliases.push(drill_alias.clone());
-
-        rate_sql = format!("{}{} as {}, ",
-            rate_sql, drill.primary_key.clone(), drill_alias
-        );
-    }
-
-    rate_sql = format!("{}{} as rate_col_id, {} as rate_mea from {}) as s1",
-        rate_sql, rate.column, meas[0].column, rate.table.name
-    );
-
-    rate_sql = format!("{} all inner join ({}) as s2", rate_sql, final_sql);
-
-    rate_sql = format!("select {}, rate_col_id, rate_mea, final_m0 from {} using {}",
-        final_drill_cols,
-        rate_sql,
-        join(drill_aliases, ", ")
-    );
-
-    rate_sql = format!("select {}, final_m0, groupArray(rate_col_id) as rate_col_id, groupArray(rate_mea) as rate_mea from ({}) group by {}, final_m0",
-        final_drill_cols,
-        rate_sql,
-        final_drill_cols
-    );
-
-    rate_sql = format!("select {}, rate_col_id_final, rate_mea_final, final_m0 from ({}) array join rate_col_id as rate_col_id_final, rate_mea as rate_mea_final",
-        final_drill_cols, rate_sql
-    );
-
-    rate_sql = format!("select {}, final_m0, sum(rate_mea_final) / avg(final_m0) from ({}) where rate_col_id_final in ({}) group by {}, final_m0 order by {}",
-        final_drill_cols,
-        rate_sql,
-        join(rate.members.clone(), ", "),
-        final_drill_cols,
-        final_drill_cols
-    );
-
-    rate_sql
-}
-
 
 
 // TODO test having not cuts or drilldowns
